refactor: restructure the codebase to be more modular
